// src/interfaces/IData.ts
interface IData {
	Date: string;
	Day: string;
	Time: string;
	Fajr: string;
	Sunrise: string;
	Zuhr: string;
	Khutbah: string;
	Asr: string;
	Maghrib: string;
	Isha: string;
	"Fajr J": string;
	"Zuhr J": string;
	"Asr J": string;
	"Maghrib J": string;
	"Isha J": string;
	"Khutbah J": string;
}

export default IData;